requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
yt-dlp==2024.4.9
ffmpeg-python==0.2.0
mutagen==1.47.0
